App Specific Functions and Internals

These functions are needed by OpenCPU apps in the package, or needed to be exported by the package but not intended to be used by users.

app_read_csv(...)
app_test(...)

.load_species_data(species, boot=NULL, path=NULL, version=NULL,
    taxon, model_north, model_south)
.calculate_results(y, level=0.9, .c4is)
.calculate_limit(y, limit=NULL)
.report_all_by1(boot=NULL, path=NULL, version=NULL, level=0.9, cores=NULL)
.read_raster_template()
.make_raster(value, rc, rt)
.rasterize_multi(y, type=c("richness", "intactness"), rt)
.verbose()
.get_cores(cores=NULL)
.check(x, ref)
.combine_veg_soil(w, veg, soil)
.validate_id(id, type=c("km", "qs"))
.verbose()
.select_id(mregion="both", nr=NULL, nsr=NULL, luf=NULL)
p_bird(D, area=c("ha", "km"), pair_adj=2)
.truncate(x, trunc=NULL)

Arguments

species, boot, path, version
arguments passed to load_species_data and other functions.

taxon, model_north, model_south
taxonomic group required for the correct path, and logical values indicating north and south model results.

y, level
arguments passed to calculate_results and other functions.

.c4is
the subset environment as.list(.c4is).

cores
desired number of cores to use.

value, rc, rt
call value, rc is data frame with Row and Col indices for value, rt is raster template from .read_raster_template.

id
spatial IDs.

type
type of multi-species map (richness or intactness) or spatial ID (km or QS).

limit
abundance threshold for multi-species intactness reporting.

x
input object.

ref
checking the validity of land cover classes in x against the reference list of possible values.

w, veg, soil
calculates a weighted average of veg and soil based models based on weights for the vegetation models (w).

mregion, nr, nsr, luf
regions.

D, area, pair_adj
density, scale, and pair adjustment to turn bird densities into probability of observing non-zero count.

trunc
quantile for truncating values of x, usually a vector or a raster object.

...
arguments passed to underlying functions.

Details

app_read_csv wraps read.csv.

app_test mimics custom_report.

Interface for internal functions might change and usage is not recommended.